# improve: move all onKill events to registed onDeath