--- a/data-otservbr-global/monster/bosses/diseased_bill.lua
+++ b/data-otservbr-global/monster/bosses/diseased_bill.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"DiseasedTrioDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 485,
 	bossRace = RARITY_BANE,