--- a/data-otservbr-global/monster/bosses/diseased_fred.lua
+++ b/data-otservbr-global/monster/bosses/diseased_fred.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"DiseasedTrioDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 484,
 	bossRace = RARITY_BANE,