--- a/data-otservbr-global/monster/bosses/raging_mage.lua
+++ b/data-otservbr-global/monster/bosses/raging_mage.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"RagingMageDeath",
+}
+
 monster.health = 3500
 monster.maxHealth = 3500
 monster.race = "blood"