--- a/data-otservbr-global/monster/bosses/splasher.lua
+++ b/data-otservbr-global/monster/bosses/splasher.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"QuaraLeadersDeath",
+}
+
 monster.health = 1700
 monster.maxHealth = 1700
 monster.race = "blood"