--- a/data-otservbr-global/monster/constructs/golden_servant_replica.lua
+++ b/data-otservbr-global/monster/constructs/golden_servant_replica.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"ReplicaServantDeath",
+}
+
 monster.raceId = 1327
 monster.Bestiary = {
 	class = "Construct",