--- a/data-otservbr-global/monster/dawnport/muglex_clan_footman.lua
+++ b/data-otservbr-global/monster/dawnport/muglex_clan_footman.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"MorrisGoblinDeath",
+}
+
 monster.health = 50
 monster.maxHealth = 50
 monster.race = "blood"