--- a/data-otservbr-global/monster/dragons/dragon.lua
+++ b/data-otservbr-global/monster/dragons/dragon.lua
@@ -13,6 +13,11 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"TheGreatDragonHuntDeath",
+	"TheFirstDragonDragonTaskDeath",
+}
+
 monster.raceId = 34
 monster.Bestiary = {
 	class = "Dragon",