--- a/data-otservbr-global/monster/dragons/dragon_lord.lua
+++ b/data-otservbr-global/monster/dragons/dragon_lord.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"TheGreatDragonHuntDeath",
+}
+
 monster.raceId = 39
 monster.Bestiary = {
 	class = "Dragon",