--- a/data-otservbr-global/monster/extra_dimensional/yielothax.lua
+++ b/data-otservbr-global/monster/extra_dimensional/yielothax.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"YielothaxDeath",
+}
+
 monster.raceId = 717
 monster.Bestiary = {
 	class = "Extra Dimensional",