--- a/data-otservbr-global/monster/humanoids/minotaur_bruiser.lua
+++ b/data-otservbr-global/monster/humanoids/minotaur_bruiser.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"MorrisMinotaurDeath",
+}
+
 monster.health = 100
 monster.maxHealth = 100
 monster.race = "blood"