--- a/data-otservbr-global/monster/humanoids/minotaur_cult_follower.lua
+++ b/data-otservbr-global/monster/humanoids/minotaur_cult_follower.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"MinotaurCultTaskDeath",
+}
+
 monster.raceId = 1508
 monster.Bestiary = {
 	class = "Humanoid",