--- a/data-otservbr-global/monster/humanoids/minotaur_cult_prophet.lua
+++ b/data-otservbr-global/monster/humanoids/minotaur_cult_prophet.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"MinotaurCultTaskDeath",
+}
+
 monster.raceId = 1509
 monster.Bestiary = {
 	class = "Humanoid",