--- a/data-otservbr-global/monster/humanoids/minotaur_cult_zealot.lua
+++ b/data-otservbr-global/monster/humanoids/minotaur_cult_zealot.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"MinotaurCultTaskDeath",
+}
+
 monster.raceId = 1510
 monster.Bestiary = {
 	class = "Humanoid",