--- a/data-otservbr-global/monster/humans/burning_gladiator.lua
+++ b/data-otservbr-global/monster/humans/burning_gladiator.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"FafnarMissionsDeath",
+}
+
 monster.raceId = 1798
 monster.Bestiary = {
 	class = "Human",