--- a/data-otservbr-global/monster/humans/nomad_blue.lua
+++ b/data-otservbr-global/monster/humans/nomad_blue.lua
@@ -14,6 +14,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"NomadDeath",
+}
+
 monster.raceId = 777
 monster.Bestiary = {
 	class = "Human",