--- a/data-otservbr-global/monster/humans/nomad_female.lua
+++ b/data-otservbr-global/monster/humans/nomad_female.lua
@@ -14,6 +14,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"NomadDeath",
+}
+
 monster.raceId = 776
 monster.Bestiary = {
 	class = "Human",