--- a/data-otservbr-global/monster/magicals/enfeebled_silencer.lua
+++ b/data-otservbr-global/monster/magicals/enfeebled_silencer.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"ThreatenedDreamsNightmareMonstersDeath",
+}
+
 monster.raceId = 1443
 monster.Bestiary = {
 	class = "Magical",