--- a/data-otservbr-global/monster/magicals/frazzlemaw.lua
+++ b/data-otservbr-global/monster/magicals/frazzlemaw.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"RoshamuulKillsDeath",
+}
+
 monster.raceId = 1022
 monster.Bestiary = {
 	class = "Magical",