--- a/data-otservbr-global/monster/magicals/silencer.lua
+++ b/data-otservbr-global/monster/magicals/silencer.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"RoshamuulKillsDeath",
+}
+
 monster.raceId = 1014
 monster.Bestiary = {
 	class = "Magical",