--- a/data-otservbr-global/monster/magicals/weakened_frazzlemaw.lua
+++ b/data-otservbr-global/monster/magicals/weakened_frazzlemaw.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"ThreatenedDreamsNightmareMonstersDeath",
+}
+
 monster.raceId = 1442
 monster.Bestiary = {
 	class = "Magical",