--- a/data-otservbr-global/monster/quests/bigfoots_burden/bosses/gnomevil.lua
+++ b/data-otservbr-global/monster/quests/bigfoots_burden/bosses/gnomevil.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"BossesWarzoneDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 893,
 	bossRace = RARITY_ARCHFOE,