--- a/data-otservbr-global/monster/quests/bigfoots_burden/versperoth.lua
+++ b/data-otservbr-global/monster/quests/bigfoots_burden/versperoth.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"VesperothDeath",
+}
+
 monster.health = 100000
 monster.maxHealth = 100000
 monster.race = "venom"