--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/dorokoll_the_mystic.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/dorokoll_the_mystic.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"EssenceOfMaliceSpawnsDeath",
+}
+
 monster.health = 30000
 monster.maxHealth = 30000
 monster.race = "blood"