--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/eshtaba_the_conjurer.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/eshtaba_the_conjurer.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"EssenceOfMaliceSpawnsDeath",
+}
+
 monster.health = 30000
 monster.maxHealth = 30000
 monster.race = "blood"