--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/mezlon_the_defiler.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/mezlon_the_defiler.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"EssenceOfMaliceSpawnsDeath",
+}
+
 monster.health = 30000
 monster.maxHealth = 30000
 monster.race = "blood"