--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/pillars/pillar_of_protection.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/pillars/pillar_of_protection.lua
@@ -14,6 +14,10 @@ monster.corpse = 0
 monster.speed = 0
 monster.manaCost = 0
 
+monster.events = {
+	"DestroyedPillar",
+}
+
 monster.changeTarget = {
 	interval = 4000,
 	chance = 20,