--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/ravenous_hunger.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/ravenous_hunger.lua
@@ -13,6 +13,11 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"LeidenHeal",
+	"CultsOfTibiaBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1427,
 	bossRace = RARITY_ARCHFOE,
@@ -55,10 +60,6 @@ monster.flags = {
 	canWalkOnEnergy = true,
 	canWalkOnFire = true,
 	canWalkOnPoison = true,
-}
-
-monster.events = {
-	"LeidenHeal",
 }
 
 monster.light = {