--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/the_false_god.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/the_false_god.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"CultsOfTibiaBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1409,
 	bossRace = RARITY_ARCHFOE,