--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/the_sandking.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/the_sandking.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"CultsOfTibiaBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1444,
 	bossRace = RARITY_ARCHFOE,