--- a/data-otservbr-global/monster/quests/cults_of_tibia/bosses/the_source_of_corruption.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/bosses/the_source_of_corruption.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"CultsOfTibiaBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1500,
 	bossRace = RARITY_ARCHFOE,