--- a/data-otservbr-global/monster/quests/cults_of_tibia/cult_enforcer.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/cult_enforcer.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"CarlinVortexDeath",
+}
+
 monster.raceId = 1513
 monster.Bestiary = {
 	class = "Human",