--- a/data-otservbr-global/monster/quests/cults_of_tibia/cult_scholar.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/cult_scholar.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"CarlinVortexDeath",
+}
+
 monster.raceId = 1514
 monster.Bestiary = {
 	class = "Human",