--- a/data-otservbr-global/monster/quests/cults_of_tibia/misguided_shadow.lua
+++ b/data-otservbr-global/monster/quests/cults_of_tibia/misguided_shadow.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"GlowingRubbishAmuletDeath",
+}
+
 monster.health = 3000
 monster.maxHealth = 3000
 monster.race = "blood"