--- a/data-otservbr-global/monster/quests/dangerous_depth/bosses/the_baron_from_below.lua
+++ b/data-otservbr-global/monster/quests/dangerous_depth/bosses/the_baron_from_below.lua
@@ -13,6 +13,11 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"DepthWarzoneBossDeath",
+	"TheBaronFromBelowThink",
+}
+
 monster.health = 350000
 monster.maxHealth = 350000
 monster.race = "blood"