--- a/data-otservbr-global/monster/quests/dangerous_depth/snail_slime.lua
+++ b/data-otservbr-global/monster/quests/dangerous_depth/snail_slime.lua
@@ -13,6 +13,11 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"SnailSlimeDeath",
+	"SnailSlimeThink",
+}
+
 monster.health = 4500
 monster.maxHealth = 4500
 monster.race = "venom"