--- a/data-otservbr-global/monster/quests/dark_trails/death_priest_shargon.lua
+++ b/data-otservbr-global/monster/quests/dark_trails/death_priest_shargon.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"ShargonDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1047,
 	bossRace = RARITY_BANE,