--- a/data-otservbr-global/monster/quests/feaster_of_souls/the_dread_maiden.lua
+++ b/data-otservbr-global/monster/quests/feaster_of_souls/the_dread_maiden.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"FeasterOfSoulsBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1872,
 	bossRace = RARITY_ARCHFOE,