--- a/data-otservbr-global/monster/quests/feaster_of_souls/the_pale_worm.lua
+++ b/data-otservbr-global/monster/quests/feaster_of_souls/the_pale_worm.lua
@@ -22,6 +22,7 @@ monster.manaCost = 0
 
 monster.events = {
 	"paleWormDeath",
+	"FeasterOfSoulsBossDeath",
 }
 
 monster.changeTarget = {