--- a/data-otservbr-global/monster/quests/feaster_of_souls/the_unwelcome.lua
+++ b/data-otservbr-global/monster/quests/feaster_of_souls/the_unwelcome.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"FeasterOfSoulsBossDeath",
+}
+
 monster.health = 300000
 monster.maxHealth = 300000
 monster.race = "undead"