--- a/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/mazoran.lua
+++ b/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/mazoran.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"AscendantBossesDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1186,
 	bossRace = RARITY_ARCHFOE,