--- a/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/plagirath.lua
+++ b/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/plagirath.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"AscendantBossesDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1199,
 	bossRace = RARITY_ARCHFOE,