--- a/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/tarbaz.lua
+++ b/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/tarbaz.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"AscendantBossesDeath",
+}
+
 monster.health = 290000
 monster.maxHealth = 290000
 monster.race = "undead"