--- a/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/the_shatterer.lua
+++ b/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/the_shatterer.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"TheShattererDeath",
+}
+
 monster.health = 220000
 monster.maxHealth = 220000
 monster.race = "fire"