--- a/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/zamulosh.lua
+++ b/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/zamulosh.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"AscendantBossesDeath",
+}
+
 monster.health = 300000
 monster.maxHealth = 300000
 monster.race = "undead"