--- a/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/zamulosh2.lua
+++ b/data-otservbr-global/monster/quests/ferumbras_ascendant/bosses/zamulosh2.lua
@@ -14,6 +14,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"AscendantBossesDeath",
+}
+
 monster.health = 300000
 monster.maxHealth = 300000
 monster.race = "undead"