--- a/data-otservbr-global/monster/quests/forgotten_knowledge/bosses/melting_frozen_horror.lua
+++ b/data-otservbr-global/monster/quests/forgotten_knowledge/bosses/melting_frozen_horror.lua
@@ -13,6 +13,11 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"ForgottenKnowledgeBossDeath",
+	"MeltingDeath",
+}
+
 monster.health = 70000
 monster.maxHealth = 70000
 monster.race = "undead"
@@ -55,10 +60,6 @@ monster.flags = {
 	canWalkOnEnergy = true,
 	canWalkOnFire = true,
 	canWalkOnPoison = true,
-}
-
-monster.events = {
-	"MeltingDeath",
 }
 
 monster.light = {