--- a/data-otservbr-global/monster/quests/forgotten_knowledge/bound_astral_power.lua
+++ b/data-otservbr-global/monster/quests/forgotten_knowledge/bound_astral_power.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"BoundAstralPowerDeath",
+}
+
 monster.health = 70000
 monster.maxHealth = 70000
 monster.race = "venom"