--- a/data-otservbr-global/monster/quests/forgotten_knowledge/cosmic_energy_prism_c.lua
+++ b/data-otservbr-global/monster/quests/forgotten_knowledge/cosmic_energy_prism_c.lua
@@ -44,7 +44,7 @@ monster.flags = {
 }
 
 monster.events = {
-	"EnergyPrism",
+	"EnergyPrismHealthChange",
 }
 
 monster.light = {