--- a/data-otservbr-global/monster/quests/grave_danger/bosses/king_zelos.lua
+++ b/data-otservbr-global/monster/quests/grave_danger/bosses/king_zelos.lua
@@ -14,6 +14,7 @@ monster.outfit = {
 }
 
 monster.events = {
+	"GraveDangerBossDeath",
 	"zelosDeath",
 }