--- a/data-otservbr-global/monster/quests/in_service_of_yalahar/azerus.lua
+++ b/data-otservbr-global/monster/quests/in_service_of_yalahar/azerus.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"AzerusDeath",
+}
+
 monster.health = 26000
 monster.maxHealth = 26000
 monster.race = "blood"