--- a/data-otservbr-global/monster/quests/in_service_of_yalahar/inky.lua
+++ b/data-otservbr-global/monster/quests/in_service_of_yalahar/inky.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"QuaraLeadersDeath",
+}
+
 monster.health = 750
 monster.maxHealth = 750
 monster.race = "blood"