--- a/data-otservbr-global/monster/quests/in_service_of_yalahar/sharptooth.lua
+++ b/data-otservbr-global/monster/quests/in_service_of_yalahar/sharptooth.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"QuaraLeadersDeath",
+}
+
 monster.health = 3100
 monster.maxHealth = 3100
 monster.race = "blood"