--- a/data-otservbr-global/monster/quests/kilmaresh/bragrumol.lua
+++ b/data-otservbr-global/monster/quests/kilmaresh/bragrumol.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"BragrumolDeath",
+}
+
 monster.health = 38000
 monster.maxHealth = 38000
 monster.race = "fire"