--- a/data-otservbr-global/monster/quests/liquid_black/jaul.lua
+++ b/data-otservbr-global/monster/quests/liquid_black/jaul.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"DeeplingBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 773,
 	bossRace = RARITY_BANE,