--- a/data-otservbr-global/monster/quests/primal_ordeal_quest/magma_bubble.lua
+++ b/data-otservbr-global/monster/quests/primal_ordeal_quest/magma_bubble.lua
@@ -15,6 +15,7 @@ monster.outfit = {
 
 monster.events = {
 	"MagmaBubbleDeath",
+	"ThePrimeOrdealBossDeath",
 }
 
 monster.bosstiary = {