--- a/data-otservbr-global/monster/quests/primal_ordeal_quest/the_primal_menace.lua
+++ b/data-otservbr-global/monster/quests/primal_ordeal_quest/the_primal_menace.lua
@@ -66,6 +66,7 @@ monster.outfit = {
 
 monster.events = {
 	"ThePrimalMenaceDeath",
+	"ThePrimeOrdealBossDeath",
 }
 
 monster.health = 400000