--- a/data-otservbr-global/monster/quests/the_dream_courts/bosses/izcandar_the_banished.lua
+++ b/data-otservbr-global/monster/quests/the_dream_courts/bosses/izcandar_the_banished.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"DreamCourtsBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 1699,
 	bossRace = RARITY_NEMESIS,