--- a/data-otservbr-global/monster/quests/the_dream_courts/bosses/the_nightmare_beast.lua
+++ b/data-otservbr-global/monster/quests/the_dream_courts/bosses/the_nightmare_beast.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"DreamCourtsBossDeath",
+}
+
 monster.health = 850000
 monster.maxHealth = 850000
 monster.race = "blood"