--- a/data-otservbr-global/monster/quests/the_elemental_spheres/fire_overlord.lua
+++ b/data-otservbr-global/monster/quests/the_elemental_spheres/fire_overlord.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"ElementalOverlordDeath",
+}
+
 monster.health = 4000
 monster.maxHealth = 4000
 monster.race = "fire"