--- a/data-otservbr-global/monster/quests/the_elemental_spheres/lord_of_the_elements.lua
+++ b/data-otservbr-global/monster/quests/the_elemental_spheres/lord_of_the_elements.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"ElementalOverlordDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 454,
 	bossRace = RARITY_ARCHFOE,