--- a/data-otservbr-global/monster/quests/the_inquisition/annihilon.lua
+++ b/data-otservbr-global/monster/quests/the_inquisition/annihilon.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"InquisitionBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 418,
 	bossRace = RARITY_BANE,