--- a/data-otservbr-global/monster/quests/the_inquisition/golgordan.lua
+++ b/data-otservbr-global/monster/quests/the_inquisition/golgordan.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"InquisitionBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 416,
 	bossRace = RARITY_BANE,