--- a/data-otservbr-global/monster/quests/the_inquisition/hellgorak.lua
+++ b/data-otservbr-global/monster/quests/the_inquisition/hellgorak.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"InquisitionBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 403,
 	bossRace = RARITY_BANE,