--- a/data-otservbr-global/monster/quests/the_inquisition/latrivan.lua
+++ b/data-otservbr-global/monster/quests/the_inquisition/latrivan.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"InquisitionBossDeath",
+}
+
 monster.health = 25000
 monster.maxHealth = 25000
 monster.race = "fire"