--- a/data-otservbr-global/monster/quests/the_inquisition/madareth.lua
+++ b/data-otservbr-global/monster/quests/the_inquisition/madareth.lua
@@ -13,6 +13,10 @@ monster.outfit = {
 	lookMount = 0,
 }
 
+monster.events = {
+	"InquisitionBossDeath",
+}
+
 monster.bosstiary = {
 	bossRaceId = 414,
 	bossRace = RARITY_BANE,