… data-otservbr-global/monster/quests/the_inquisition/ungreez.lua
… data-otservbr-global/monster/quests/the_inquisition/ushuriel.lua
… data-otservbr-global/monster/quests/the_inquisition/zugurosh.lua
… data-otservbr-global/monster/quests/the_new_frontier/shard_of_corruption.lua
… data-otservbr-global/monster/quests/the_new_frontier/tirecz.lua
… data-otservbr-global/monster/quests/the_secret_library/bosses/ghulosh.lua
… data-otservbr-global/monster/quests/the_secret_library/bosses/gorzindel.lua
… data-otservbr-global/monster/quests/the_secret_library/bosses/the_scourge_of_oblivion.lua
… data-otservbr-global/monster/quests/the_secret_library/lokathmor.lua
… data-otservbr-global/monster/quests/the_secret_library/mazzinor.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/fury_of_the_emperor.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/lizard_abomination.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/mutated_zalamon.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/scorn_of_the_emperor.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/snake_god_essence.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/snake_thing.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/spite_of_the_emperor.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/the_keeper.lua
… data-otservbr-global/monster/quests/wrath_of_the_emperor/wrath_of_the_emperor.lua
… data-otservbr-global/monster/reptiles/lizard_magistratus.lua
… data-otservbr-global/monster/reptiles/lizard_noble.lua
… data-otservbr-global/monster/undeads/demon_skeleton.lua
… data-otservbr-global/monster/vermins/deepworm.lua
… data-otservbr-global/monster/vermins/diremaw.lua
… data-otservbr-global/monster/vermins/drillworm.lua
… data-otservbr-global/monster/vermins/parasite.lua
… data-otservbr-global/monster/vermins/wiggler.lua
… data-otservbr-global/scripts/actions/quests/dangerous_depth/using_crystals.lua
… data-otservbr-global/scripts/actions/quests/feaster_of_souls/bosses_killed.lua
… data-otservbr-global/scripts/actions/quests/grave_danger/bosses_killed.lua