Alternative implementation of #1665

This change deprecates `onKill` (you'll get a warning if you try to use `onKill`, but it'll keep working until we decide to delete the callback from the source. Taking advantage of registed `onDeath` events and the fact that mechanics often use either damage map or the party of the top damage dealer.

It also fixes a lot of inconsistencies in how kill tasks work, making most systems match what they should do when doing quests in a group.

Full credit to @duuh30 for finding the issue and exploring the fix, and @beats-dh for helping with that.

## (very basic) benchmarks

### With this change (note, also includes #1757)

```
Average (Mean): Approximately 0.353 ms
Median: Approximately 0.509 ms
95th Percentile (P95): 0.675 ms
```

### Before this change

```
Average (Mean): Approximately 2998.6 ms
Median: Approximately 2012.4 ms
95th Percentile (P95): 5997.9 ms
```

Co-Authored-By: duuh30 <duuh30@users.noreply.github.com>
Co-Authored-By: beats-dh <beats-dh@users.noreply.github.com>